+++ packages/xchain-bitcoin/src/client.ts
@@ -47,6 +47,7 @@ class Client extends UTXOClient {
     rootDerivationPaths = {
       [Network.Mainnet]: `84'/0'/0'/0/`, //note this isn't bip44 compliant, but it keeps the wallets generated compatible to pre HD wallets
       [Network.Testnet]: `84'/1'/0'/0/`,
+      [Network.Stagenet]: `84'/1'/0'/0/`, // stagenet is not configured, use testnet value
     },
     phrase = '',
   }: BitcoinClientParams) {
@@ -85,6 +86,8 @@ class Client extends UTXOClient {
       case Network.Mainnet:
         return 'https://blockstream.info'
       case Network.Testnet:
+        return 'https://blockstream.info/testnet'
+      case Network.Stagenet: // stagenet is not configured, use testnet value
         return 'https://blockstream.info/testnet'
     }
   }