+++ packages/xchain-polkadot/src/client.ts
@@ -54,6 +54,7 @@ class Client implements PolkadotClient, XChainClient {
     rootDerivationPaths = {
       [Network.Mainnet]: "44//354//0//0//0'", //TODO IS the root path we want to use?
       [Network.Testnet]: "44//354//0//0//0'",
+      [Network.Stagenet]: "44//354//0//0//0'", // stagenet is not configured, use testnet value
     },
   }: XChainClientParams) {
     this.network = network
@@ -124,6 +125,8 @@ class Client implements PolkadotClient, XChainClient {
         return 'https://polkadot.subscan.io'
       case Network.Testnet:
         return 'https://westend.subscan.io'
+      case Network.Stagenet: // stagenet is not configured, use testnet value
+        return 'https://westend.subscan.io'
     }
   }
 
@@ -138,6 +141,8 @@ class Client implements PolkadotClient, XChainClient {
         return 'wss://rpc.polkadot.io'
       case Network.Testnet:
         return 'wss://westend-rpc.polkadot.io'
+      case Network.Stagenet: // stagenet is not configured, use testnet value
+        return 'wss://westend-rpc.polkadot.io'
     }
   }
 
@@ -151,6 +156,8 @@ class Client implements PolkadotClient, XChainClient {
       case Network.Mainnet:
         return 'https://polkadot.subscan.io'
       case Network.Testnet:
+        return 'https://westend.subscan.io'
+      case Network.Stagenet: // stagenet is not configured, use testnet value
         return 'https://westend.subscan.io'
     }
   }
@@ -185,6 +192,8 @@ class Client implements PolkadotClient, XChainClient {
       case Network.Mainnet:
         return 0
       case Network.Testnet:
+        return 42
+      case Network.Stagenet: // stagenet is not configured, use testnet value
         return 42
     }
   }